backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.19
pydantic-settings==2.3.4
python-dotenv==1.0.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
alembic==1.13.1
cloudinary==1.40.0
sib-api-v3-sdk==7.6.0
httpx==0.27.0
PyPDf2==3.0.1
reportlab==4.2.0
